+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Upcoming <!-- Add unreleased change notes here: -->
+- Added StagedEvent.save_proto! shortcut method with documentation
 
 ## v0.0.1
 *Release Date: 10/12/2021*